[fix]: Build error suspense boundary
